Performance Updates

+++ source_files/edge/i_system.cc
@@ -141,9 +141,9 @@ int PureRandomNumber(void)
     return (P1 ^ P2) & 0x7FFFFFFF;
 }
 
-uint32_t GetMicroseconds(void)
+uint64_t GetMicroseconds(void)
 {
-    return (uint32_t)std::chrono::duration_cast<std::chrono::microseconds>(
+    return (uint64_t)std::chrono::duration_cast<std::chrono::microseconds>(
                std::chrono::system_clock::now().time_since_epoch())
         .count();
 }
