+++ source_files/edge/i_system.h
@@ -112,9 +112,8 @@ int GetTime(void);
 int GetMilliseconds(void);
 
 // Returns a value that increases by 1000000 every second (i.e. each unit
-// is a single microsecond).  Since this value will wrap-around regularly
-// (roughly every 71 minutes), caller *MUST* check for this situation.
-uint32_t GetMicroseconds(void);
+// is a single microsecond).
+uint64_t GetMicroseconds(void);
 
 //--------------------------------------------------------
 //  MUSIC functions.
